dash/requirements.txt
dash==3.0.0
dash-mantine-components==1.1.0
dash-chat==0.2.0
requests==2.32.3
python-dotenv==1.0.0
